change integer to bigint

+++ db/migrate/20140828153845_create_provinces.rb
@@ -7,11 +7,11 @@ class CreateProvinces < ActiveRecord::Migration
       t.string :latitude
       t.string :longitude
       t.string :capital
-      t.integer :area_km_sq
-      t.integer :population
+      t.bigint :area_km_sq
+      t.bigint :population
       t.integer :population_density
-      t.integer :gdp_cny
-      t.integer :gdp_usd
+      t.bigint :gdp_cny
+      t.bigint :gdp_usd
       t.integer :gdp_per_capita
       t.string :jvector_code
